+++ libs/common/src/lib/types/order-with-account.type.ts
@@ -1,5 +1,8 @@
-import { Account, Order, Platform } from '@prisma/client';
+import { Account, Order, Platform, SymbolProfile } from '@prisma/client';
 
 type AccountWithPlatform = Account & { Platform?: Platform };
 
-export type OrderWithAccount = Order & { Account?: AccountWithPlatform };
+export type OrderWithAccount = Order & {
+  Account?: AccountWithPlatform;
+  SymbolProfile?: SymbolProfile;
+};